Add black; Move to twitter client logic for interact with twitter. Make run into sep logic to expand later
