+++ Pipfile
@@ -15,6 +15,7 @@ gunicorn = "*"
 [dev-packages]
 ipython = "*"
 ipdb = "*"
+black = "*"
 
 [requires]
 python_version = "3.13"
