+++ process_tweet.py
@@ -1,76 +1,14 @@
-import json
 import logging
-import os
 from atproto import Client
-from playwright.sync_api import sync_playwright
 import requests
+from twitter_client import get_tweet_data
 
 CREDS_PATH = "credentials.json"
-HEADLESS = True
 
 logging.basicConfig(
     level=logging.INFO,
     format='%(asctime)s - %(levelname)s - %(message)s'
 )
-
-
-def get_tweet_data(url: str) -> dict:
-    """
-    Scrape a single tweet page for Tweet data
-    """
-    logging.info(f"Starting to scrape tweet from URL: {url}")
-
-    _xhr_calls = []
-
-    def intercept_response(response):
-        """capture all background requests and save them"""
-        # we can extract details from background requests
-        if response.request.resource_type == "xhr":
-            _xhr_calls.append(response)
-        return response
-
-    with sync_playwright() as pw:
-        browser = pw.chromium.launch(headless=HEADLESS)
-        context = browser.new_context(
-            user_agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/113.0.0.0 Safari/537.36",
-            viewport={"width": 1920, "height": 1080}
-        )
-        page = context.new_page()
-
-
-        # enable background request intercepting:
-        page.on("response", intercept_response)
-        # go to url and wait for the page to load
-        page.goto(url)
-        page.wait_for_selector("[data-testid='tweet']")
-
-        # find all tweet background requests:
-        tweet_calls = [f for f in _xhr_calls if "TweetResultByRestId" in f.url]
-        for xhr in tweet_calls:
-            data = xhr.json()
-            logging.info("Successfully scraped tweet data.")
-            return data['data']['tweetResult']['result']
-    
-    logging.warning("No tweet data found.")
-    raise Exception("No tweet data found.")
-
-def _clean_tweet_text(full_text: str) -> str:
-    """
-    Clean tweet text by removing t.co URLs
-    
-    Args:
-        full_text (str): Raw tweet text from Twitter API
-    
-    Returns:
-        str: Cleaned tweet text without t.co URLs
-    """
-    # Regular expression to match t.co URLs
-    import re
-    
-    # Remove t.co URLs
-    cleaned_text = re.sub(r'https://t\.co/\w+', '', full_text).strip()
-    
-    return cleaned_text
 
 def post_to_bluesky(tweet_url: str, username: str, password: str):
     """
@@ -134,3 +72,21 @@ def post_to_bluesky(tweet_url: str, username: str, password: str):
     else:
         logging.warning("Failed to process any media. Posting text only.")
         return client.send_post(tweet_text)
+
+def _clean_tweet_text(full_text: str) -> str:
+    """
+    Clean tweet text by removing t.co URLs
+    
+    Args:
+        full_text (str): Raw tweet text from Twitter API
+    
+    Returns:
+        str: Cleaned tweet text without t.co URLs
+    """
+    # Regular expression to match t.co URLs
+    import re
+    
+    # Remove t.co URLs
+    cleaned_text = re.sub(r'https://t\.co/\w+', '', full_text).strip()
+    
+    return cleaned_text
